requirements.txt
appdirs==1.4.4
blinker==1.6.2
build==0.10.0
buildozer @ https://github.com/kivy/buildozer/archive/master.zip#sha256=c96522cf32e2488bb25cf66d82fb6e1d520b678cfdf6d6c0c80ea548140abbf2
click==8.1.3
colorama==0.4.6
Cython==0.29.35
distlib==0.3.6
filelock==3.12.1
Flask==2.3.2
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
packaging==23.1
pep517==0.13.0
pexpect==4.8.0
platformdirs==3.5.3
ptyprocess==0.7.0
pyproject_hooks==1.0.0
python-for-android==2023.5.21
sh==1.14.3
toml==0.10.2
tomli==2.0.1
unzip==1.0.0
virtualenv==20.23.0
Werkzeug==2.3.6
